// generated/models/UserInformations.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserInformations = {
    id: string;
    userId: string;
    address: string;
    city: string;
    firstname: string;
    lastname: string;
    phoneNumber: string;
    siret: string;
    state: string;
    tva: number;
    zip: string;
    email: string;
};